modify ShoppingCartService.GetPotterSerialTotalPrice() logic and return type to decimal by added pair promotion price, fixed second test case

+++ PotterShoppingCart.Tests/ShoppingCartTest.cs
@@ -37,7 +37,7 @@ namespace PotterShoppingCart.Tests
             int expected = 100;
 
             var target = new ShoppingCartService();
-            int actual = target.GetPotterSerialTotalPrice(order);
+            decimal actual = target.GetPotterSerialTotalPrice(order);
 
             Assert.AreEqual(expected, actual);
         }
@@ -52,7 +52,7 @@ namespace PotterShoppingCart.Tests
             int expected = 190;
 
             var target = new ShoppingCartService();
-            int actual = target.GetPotterSerialTotalPrice(order);
+            decimal actual = target.GetPotterSerialTotalPrice(order);
 
             Assert.AreEqual(expected, actual);
         }
